Fix race in fake to real chunk replacement with Sodium (fixes #19)
If a fake chunk is queued for loading but has not yet finished loading by the
time a real chunk arrives for the same position, then we need to cancel the fake
one as otherwise it will at a later point overwrite the real one.
This will hide any changes which happened since we saved the fake chunk and
break lighting updates (cause we bypass the lighting engine for fake chunks).