Support for Sodium 0.2.0
The only relevant change was the removal of the SodiumChunkManager in favor of a
Mixin which does the same thing.
Accordingly this commit moves the ChunkStatusListener tracking out of the
sodium-specific mixin (we no longer need to duplicate code cause Sodium no
longer duplicates ClientChunkManager code), abstracted behind its own interface
so the code loads even when Sodium is not installed.
Same thing with the Sodium-specific FakeChunkManager subclass.
The Sodium-specific code then merely hooks up Sodium's listener to Bobby's
generic listener functionality.