Fix sky light of entirely occluded/unoccluded chunks (fixes #24)
If one of the lighting sections (sky/block) of a chunk is entirely dark (for
blocks and sky) or completely light (for sky), then it is not actually stored on disk.
We still need to load those sections though because otherwise the lighting
engine will fallback to its default for missing chunks, which is generally not
what we want (fully lit ocean floors and dark blocks if the chunk above them is
empty).

As a nice bonus, we can do these computations ahead of time on a background thread.