Split chunk center / view distance tracking into its own class
Because we will re-use at least the unload part of it to determine which real
chunks will get unloaded on server view distance updates, so we can safe those
and replace them with fake ones.

Also makes it easier to optimize the diffing, at least in principle. I'm putting
a lot of faith into the JVM inliner.